+++ docs/source/whatsnew.rst
@@ -3,6 +3,7 @@ What's New
 
 v2.1.1
 ------
+- :meth:`wxcadm.Org.get_audit_events()` was added to support auditing of Control Hub changes
 - :meth:`wxcadm.Org.numbers` now uses the Webex for Developers API rather than CP-API
 
 v2.1.0